requirements.txt
aiogram
asyncssh
paramiko
python-dotenv
setuptools
wheel